feat(ssp): 完善广告监测功能并优化数据类型
- 添加下载监测(dm)、下载完成监测(dsm)、视频播放监测(vpm)和视频播放完成监测(vpcm)字段
- 将publisherId和siteId的数据类型从Long统一改为Integer以提高性能
- 新增日志采样配置相关的模型和事件处理功能
- 重构元数据缓存服务中的数据类型转换逻辑
- 移除已废弃的AdMetadataListener类
- 更新数据库表结构将ID字段改为INT类型以优化存储空间
- 为publisher管理接口增加搜索功能
- 优化报表统计中的数据聚合逻辑
- 添加protobuf中publisher ID字段的类型定义更新
- 集成新的日志采样配置到域事件监听器中
- 更新代码生成器中的数据库连接配置
- 重构Stream操作使用自定义工具类提高性能
- 添加缺失的函数式编程相关依赖引入

+++ open-adexchange-api/open-adexchange-ssp-api/src/main/java/top/openadexchange/openapi/ssp/application/factory/BidRequestBuilder.java
@@ -64,7 +64,6 @@ public class BidRequestBuilder {
         builder.setId(imp.getId());
         builder.setTagid(imp.getTagid());
 
-        //MetadataRepository metadataRepository = oaxEngineServices.getCachedMetadataRepository();
         SiteAdPlacementAggregate siteAdPlacementAggregate =
                 metadataCacheService.getSiteAdPlacementByTagId(imp.getTagid());
         Assert.notNull(siteAdPlacementAggregate, "未找到对应的媒体广告位");
@@ -212,7 +211,7 @@ public class BidRequestBuilder {
     private top.openadexchange.model.Site getSite(AdGetRequest request) {
         String tagId = request.getImp().getFirst().getTagid();
         SiteAdPlacementAggregate siteAdPlacement = metadataCacheService.getSiteAdPlacementByTagId(tagId);
-        Long siteId = siteAdPlacement.getSiteAdPlacement().getSiteId();
+        Integer siteId = siteAdPlacement.getSiteAdPlacement().getSiteId();
         top.openadexchange.model.Site site = metadataCacheService.getSite(siteId);
         return site;
     }
