+++ open-adexchange-dao/src/main/java/top/openadexchange/dao/impl/SiteDaoImpl.java
@@ -19,13 +19,13 @@ import top.openadexchange.model.Site;
 public class SiteDaoImpl extends ServiceImpl<SiteMapper, Site> implements SiteDao {
 
     @Override
-    public Boolean enableSite(Long id) {
+    public Boolean enableSite(Integer id) {
         Assert.notNull(id, "id cannot be null");
         return updateChain().set(Site::getStatus, 1).eq(Site::getId, id).update();
     }
 
     @Override
-    public Boolean disableSite(Long id) {
+    public Boolean disableSite(Integer id) {
         Assert.notNull(id, "id cannot be null");
         return updateChain().set(Site::getStatus, 0).eq(Site::getId, id).update();
     }
